Separate out `IPAddress` into `ip_address.py`
