+++ chia/types/peer_info.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from typing import Union
 
 from chia.util.ints import uint16, uint64
-from chia.util.network import IPAddress
+from chia.util.ip_address import IPAddress
 from chia.util.streamable import Streamable, streamable
